Split loading/parsing a config file apart from command-line args parsing.
