+++ src/config.rs
@@ -1,20 +1,16 @@
 use std::{
-    env,
     fs::{canonicalize, read_to_string},
-    io::{stderr, Write},
-    path::{Path, PathBuf},
+    path::PathBuf,
     process,
 };
 
 use crypto_mac::{InvalidKeyLength, Mac};
-use getopts::Options;
 use hmac::Hmac;
 use lrlex::lrlex_mod;
 use lrpar::{lrpar_mod, Lexeme, Lexer};
 use regex::Regex;
 use secstr::SecStr;
 use sha1::Sha1;
-use users::{get_current_uid, get_user_by_uid, os::unix::UserExt};
 
 use crate::{config_ast, fatal, fatal_err};
 
@@ -22,10 +18,6 @@ type StorageT = u8;
 
 lrlex_mod!("config.l");
 lrpar_mod!("config.y");
-
-/// Default locations to look for `snare.conf`: `~/` will be automatically converted to the current
-/// user's home directory.
-const SNARE_CONF_SEARCH: &[&str] = &["/etc/snare.conf", "~/.snare.conf"];
 
 pub struct Config {
     /// The maximum number of parallel jobs to run.
@@ -37,23 +29,8 @@ pub struct Config {
 }
 
 impl Config {
-    /// Create a `Config`.
-    pub fn new() -> Self {
-        let args: Vec<String> = env::args().collect();
-        let prog = &args[0];
-        let matches = Options::new()
-            .optmulti("c", "config", "Path to snare.conf.", "<path>")
-            .optflag("h", "help", "")
-            .parse(&args[1..])
-            .unwrap_or_else(|_| usage(prog));
-        if matches.opt_present("h") {
-            usage(prog);
-        }
-
-        let conf_path = match matches.opt_str("c") {
-            Some(p) => PathBuf::from(&p),
-            None => search_snare_conf().unwrap_or_else(|| fatal("Can't find snare.conf")),
-        };
+    /// Create a `Config` from `path`.
+    pub fn from_path(conf_path: PathBuf) -> Self {
         let input = match read_to_string(conf_path) {
             Ok(s) => s,
             Err(e) => fatal_err("Can't read configuration file", e),
@@ -273,44 +250,6 @@ fn conf_fatal(lexer: &dyn Lexer<StorageT>, lexeme: Lexeme<StorageT>, msg: &str) 
         line.trim(),
         msg
     ));
-}
-
-/// Print out program usage then exit.
-fn usage(prog: &str) -> ! {
-    let path = Path::new(prog);
-    let leaf = path
-        .file_name()
-        .map(|x| x.to_str().unwrap_or("snare"))
-        .unwrap_or("snare");
-    writeln!(
-        &mut stderr(),
-        "Usage: {} [-e email] [-j <max-jobs>] -p <port> -r <repos-dir> -s <secrets-path>",
-        leaf
-    )
-    .ok();
-    process::exit(1)
-}
-
-/// Try to find a `snare.conf` file.
-fn search_snare_conf() -> Option<PathBuf> {
-    for cnd in SNARE_CONF_SEARCH {
-        if cnd.starts_with("~/") {
-            let mut homedir =
-                match get_user_by_uid(get_current_uid()).map(|x| x.home_dir().to_path_buf()) {
-                    Some(p) => p,
-                    None => continue,
-                };
-            homedir.push(&cnd[2..]);
-            if homedir.is_file() {
-                return Some(homedir);
-            }
-        }
-        let p = PathBuf::from(cnd);
-        if p.is_file() {
-            return Some(p);
-        }
-    }
-    None
 }
 
 /// The configuration for a given repository.
