+++ src/main.rs
@@ -13,19 +13,28 @@ mod jobrunner;
 mod queue;
 
 use std::{
+    env,
     error::Error,
     fmt::Display,
+    io::{stderr, Write},
     net::SocketAddr,
     os::unix::io::RawFd,
+    path::{Path, PathBuf},
     process,
     sync::{Arc, Mutex},
 };
 
+use getopts::Options;
 use hyper::Server;
 use nix::{fcntl::OFlag, unistd::pipe2};
+use users::{get_current_uid, get_user_by_uid, os::unix::UserExt};
 
 use config::Config;
 use queue::Queue;
+
+/// Default locations to look for `snare.conf`: `~/` will be automatically converted to the current
+/// user's home directory.
+const SNARE_CONF_SEARCH: &[&str] = &["/etc/snare.conf", "~/.snare.conf"];
 
 pub(crate) struct Snare {
     config: Mutex<Config>,
@@ -46,9 +55,63 @@ fn fatal_err<E: Into<Box<dyn Error>> + Display>(msg: &str, err: E) -> ! {
     fatal(&format!("{}: {}", msg, err));
 }
 
+/// Try to find a `snare.conf` file.
+fn search_snare_conf() -> Option<PathBuf> {
+    for cnd in SNARE_CONF_SEARCH {
+        if cnd.starts_with("~/") {
+            let mut homedir =
+                match get_user_by_uid(get_current_uid()).map(|x| x.home_dir().to_path_buf()) {
+                    Some(p) => p,
+                    None => continue,
+                };
+            homedir.push(&cnd[2..]);
+            if homedir.is_file() {
+                return Some(homedir);
+            }
+        }
+        let p = PathBuf::from(cnd);
+        if p.is_file() {
+            return Some(p);
+        }
+    }
+    None
+}
+
+/// Print out program usage then exit.
+fn usage(prog: &str) -> ! {
+    let path = Path::new(prog);
+    let leaf = path
+        .file_name()
+        .map(|x| x.to_str().unwrap_or("snare"))
+        .unwrap_or("snare");
+    writeln!(
+        &mut stderr(),
+        "Usage: {} [-e email] [-j <max-jobs>] -p <port> -r <repos-dir> -s <secrets-path>",
+        leaf
+    )
+    .ok();
+    process::exit(1)
+}
+
 #[tokio::main]
 pub async fn main() {
-    let config = Config::new();
+    let args: Vec<String> = env::args().collect();
+    let prog = &args[0];
+    let matches = Options::new()
+        .optmulti("c", "config", "Path to snare.conf.", "<path>")
+        .optflag("h", "help", "")
+        .parse(&args[1..])
+        .unwrap_or_else(|_| usage(prog));
+    if matches.opt_present("h") {
+        usage(prog);
+    }
+
+    let conf_path = match matches.opt_str("c") {
+        Some(p) => PathBuf::from(&p),
+        None => search_snare_conf().unwrap_or_else(|| fatal("Can't find snare.conf")),
+    };
+    let config = Config::from_path(conf_path);
+
     let addr = SocketAddr::from(([0, 0, 0, 0], config.port));
     let server = match Server::try_bind(&addr) {
         Ok(s) => s,
